ESP32 board is triggered by 3v3 on GPIO1
Also, added esp23s3-muon-detector-led.yaml, which includes code for driving
an RGB LED effect on RGB strip, attached to GPIO2.

The LED Strip can be powered by the 5V from the ESP32S3 board, but with not to
many LEDs lit.

The 'shower' effect is automatically selected when the 'light' is turned on
(which is the configured default on boot).

The 'shower' effect continuously shifts colours to the next higher LED, so by
setting a colour in LED0, it will be pushed down the strip.